Fix "tar1090" display
When using more than one tar1090 project, services are not displayed correctly in the "check status adsb services" section. This is only a quick fix; it should be properly resolved in the info_ctl method.

Here is my solution for tar1090-adsblol and tar1090-adsbx. If you want to add another tar1090-"name" project, it must be defined here in order to display correctly.